refactor: move separationCandidatesForCycles and related code to separate file
